+++ crates/stackable-versioned/src/gen/field.rs
@@ -9,7 +9,7 @@ use syn::{Field, Ident};
 use crate::{
     attrs::field::FieldAttributes,
     consts::DEPRECATED_PREFIX,
-    gen::{version::ContainerVersion, ToTokensExt},
+    gen::{neighbors::Neighbors, version::ContainerVersion, ToTokensExt},
 };
 
 /// A versioned field, which contains contains common [`Field`] data and a chain
@@ -36,96 +36,29 @@ impl ToTokensExt for VersionedField {
                 // The code generation then depends on the relation to other
                 // versions (with actions).
 
-                // TODO (@Techassi): Make this more robust by also including
-                // the container versions in the action chain. I'm not happy
-                // with the follwoing code at all. It serves as a good first
-                // implementation to get something out of the door.
-                match chain.get(&container_version.inner) {
-                    Some(action) => match action {
-                        FieldStatus::Added(field_ident) => {
-                            let field_type = &self.inner.ty;
+                let field_type = &self.inner.ty;
 
-                            Some(quote! {
-                                pub #field_ident: #field_type,
-                            })
-                        }
-                        FieldStatus::Renamed { from: _, to } => {
-                            let field_type = &self.inner.ty;
-
-                            Some(quote! {
-                                pub #to: #field_type,
-                            })
-                        }
-                        FieldStatus::Deprecated(field_ident) => {
-                            let field_type = &self.inner.ty;
-
-                            Some(quote! {
-                                #[deprecated]
-                                pub #field_ident: #field_type,
-                            })
-                        }
-                    },
-                    None => {
-                        // Generate field if the container version is not
-                        // included in the action chain. First we check the
-                        // earliest field action version.
-                        if let Some((version, action)) = chain.first_key_value() {
-                            if container_version.inner < *version {
-                                match action {
-                                    FieldStatus::Added(_) => return None,
-                                    FieldStatus::Renamed { from, to: _ } => {
-                                        let field_type = &self.inner.ty;
-
-                                        return Some(quote! {
-                                            pub #from: #field_type,
-                                        });
-                                    }
-                                    FieldStatus::Deprecated(field_ident) => {
-                                        let field_type = &self.inner.ty;
-
-                                        return Some(quote! {
-                                            pub #field_ident: #field_type,
-                                        });
-                                    }
-                                }
-                            }
-                        }
-
-                        // Check the container version against the latest
-                        // field action version.
-                        if let Some((version, action)) = chain.last_key_value() {
-                            if container_version.inner > *version {
-                                match action {
-                                    FieldStatus::Added(field_ident) => {
-                                        let field_type = &self.inner.ty;
-
-                                        return Some(quote! {
-                                            pub #field_ident: #field_type,
-                                        });
-                                    }
-                                    FieldStatus::Renamed { from: _, to } => {
-                                        let field_type = &self.inner.ty;
-
-                                        return Some(quote! {
-                                            pub #to: #field_type,
-                                        });
-                                    }
-                                    FieldStatus::Deprecated(field_ident) => {
-                                        let field_type = &self.inner.ty;
-
-                                        return Some(quote! {
-                                            #[deprecated]
-                                            pub #field_ident: #field_type,
-                                        });
-                                    }
-                                }
-                            }
-                        }
-
-                        // TODO (@Techassi): Handle versions which are in between
-                        // versions defined in field actions.
-                        None
-                    }
+                match chain
+                    .get(&container_version.inner)
+                    .expect("internal error: chain must contain container version")
+                {
+                    FieldStatus::Added(field_ident) => Some(quote! {
+                        pub #field_ident: #field_type,
+                    }),
+                    FieldStatus::Renamed { _from: _, to } => Some(quote! {
+                        pub #to: #field_type,
+                    }),
+                    FieldStatus::Deprecated {
+                        ident: field_ident,
+                        note,
+                    } => Some(quote! {
+                        #[deprecated = #note]
+                        pub #field_ident: #field_type,
+                    }),
+                    FieldStatus::NotPresent => None,
+                    FieldStatus::NoChange(field_ident) => Some(quote! {
+                        pub #field_ident: #field_type,
+                    }),
                 }
             }
             None => {
@@ -144,11 +77,16 @@ impl ToTokensExt for VersionedField {
 }
 
 impl VersionedField {
+    /// Create a new versioned field by creating a status chain for each version
+    /// defined in an action in the field attribute.
+    ///
+    /// This chain will get extended by the versions defined on the container by
+    /// calling the [`VersionedField::insert_container_versions`] function.
     pub(crate) fn new(field: Field, attrs: FieldAttributes) -> Result<Self, Error> {
-        // Constructing the change chain requires going through the actions from
+        // Constructing the action chain requires going through the actions from
         // the end, because the base struct always represents the latest (most
         // up-to-date) version of that struct. That's why the following code
-        // needs to go through the changes in reverse order, as otherwise it is
+        // needs to go through the actions in reverse order, as otherwise it is
         // impossible to extract the field ident for each version.
 
         // Deprecating a field is always the last state a field can end up in. For
@@ -160,7 +98,13 @@ impl VersionedField {
             let mut actions = BTreeMap::new();
 
             let ident = field.ident.as_ref().unwrap();
-            actions.insert(*deprecated.since, FieldStatus::Deprecated(ident.clone()));
+            actions.insert(
+                *deprecated.since,
+                FieldStatus::Deprecated {
+                    ident: ident.clone(),
+                    note: deprecated.note.to_string(),
+                },
+            );
 
             // When the field is deprecated, any rename which occured beforehand
             // requires access to the field ident to infer the field ident for
@@ -175,7 +119,7 @@ impl VersionedField {
                 actions.insert(
                     *rename.since,
                     FieldStatus::Renamed {
-                        from: from.clone(),
+                        _from: from.clone(),
                         to: ident,
                     },
                 );
@@ -201,7 +145,7 @@ impl VersionedField {
                 actions.insert(
                     *rename.since,
                     FieldStatus::Renamed {
-                        from: from.clone(),
+                        _from: from.clone(),
                         to: ident,
                     },
                 );
@@ -241,11 +185,58 @@ impl VersionedField {
             })
         }
     }
+
+    /// Inserts container versions not yet present in the status chain.
+    ///
+    /// When intially creating a new [`VersionedField`], the code doesn't have
+    /// access to the versions defined on the container. This function inserts
+    /// all non-present container versions and decides which status and ident
+    /// is the right fit based on the status neighbors.
+    ///
+    /// This continous chain ensures that when generating code (tokens), each
+    /// field can lookup the status for a requested version.
+    pub(crate) fn insert_container_versions(&mut self, versions: &Vec<ContainerVersion>) {
+        if let Some(chain) = &mut self.chain {
+            for version in versions {
+                if chain.contains_key(&version.inner) {
+                    continue;
+                }
+
+                match chain.get_neighbors(&version.inner) {
+                    (None, Some(_)) => chain.insert(version.inner, FieldStatus::NotPresent),
+                    (Some(status), None) => {
+                        let ident = match status {
+                            FieldStatus::Added(ident) => ident,
+                            FieldStatus::Renamed { _from: _, to } => to,
+                            FieldStatus::Deprecated { ident, note: _ } => ident,
+                            FieldStatus::NoChange(ident) => ident,
+                            FieldStatus::NotPresent => unreachable!(),
+                        };
+
+                        chain.insert(version.inner, FieldStatus::NoChange(ident.clone()))
+                    }
+                    (Some(status), Some(_)) => {
+                        let ident = match status {
+                            FieldStatus::Added(ident) => ident,
+                            FieldStatus::Renamed { _from: _, to } => to,
+                            FieldStatus::NoChange(ident) => ident,
+                            _ => unreachable!(),
+                        };
+
+                        chain.insert(version.inner, FieldStatus::NoChange(ident.clone()))
+                    }
+                    _ => unreachable!(),
+                };
+            }
+        }
+    }
 }
 
 #[derive(Debug)]
 pub(crate) enum FieldStatus {
     Added(Ident),
-    Renamed { from: Ident, to: Ident },
-    Deprecated(Ident),
+    Renamed { _from: Ident, to: Ident },
+    Deprecated { ident: Ident, note: String },
+    NoChange(Ident),
+    NotPresent,
 }