+++ crates/stackable-versioned/src/gen/mod.rs
@@ -9,6 +9,7 @@ use crate::{
 };
 
 pub(crate) mod field;
+pub(crate) mod neighbors;
 pub(crate) mod venum;
 pub(crate) mod version;
 pub(crate) mod vstruct;