+++ crates/stackable-versioned/CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 
+- Improve action chain generation ([#784]).
+
+[#784](ttps://github.com/stackabletech/operator-rs/pull/784)
+
 ## [0.1.0] - 2024-05-08
 
 ### Changed